+++ src/domain/flow-system/services/tests/tests/test_flow_agent_impl.rs
@@ -15,10 +15,11 @@ use kamu_accounts::{AccountConfig, CurrentAccountSubject};
 use kamu_adapter_flow_dataset::*;
 use kamu_adapter_task_dataset::*;
 use kamu_core::{CompactionResult, PullResult, ResetResult, TransformStatus};
-use kamu_datasets::{DatasetEntryServiceExt, DatasetIntervalIncrement};
+use kamu_datasets::DatasetEntryServiceExt;
 use kamu_datasets_services::testing::MockDatasetIncrementQueryService;
 use kamu_flow_system::*;
 use kamu_task_system::*;
+use odf::dataset::MetadataChainIncrementInterval;
 
 use super::{
     FlowHarness,
@@ -901,7 +902,7 @@ async fn test_ingest_flow_with_multiple_iterations() {
     let harness = FlowHarness::with_overrides(FlowHarnessOverrides {
         mock_transform_flow_evaluator: Some(mock_transform_flow_evaluator),
         mock_dataset_changes: Some(MockDatasetIncrementQueryService::with_increment_between(
-            DatasetIntervalIncrement {
+            MetadataChainIncrementInterval {
                 num_blocks: 1,
                 num_records: 3,
                 updated_watermark: None,
@@ -976,6 +977,7 @@ async fn test_ingest_flow_with_multiple_iterations() {
                                 new_head: odf::Multihash::from_digest_sha3_256(b"newest-slice"),
                                 has_more: true,
                             },
+                            data_increment: None,
                         }
                         .into_task_result(),
                     ),
@@ -1006,6 +1008,7 @@ async fn test_ingest_flow_with_multiple_iterations() {
                                 new_head: odf::Multihash::from_digest_sha3_256(b"newest-slice-1"),
                                 has_more: true,
                             },
+                            data_increment: None,
                         }
                         .into_task_result(),
                     ),
@@ -3442,7 +3445,7 @@ async fn test_derived_dataset_triggered_after_input_change() {
     let harness = FlowHarness::with_overrides(FlowHarnessOverrides {
         mock_transform_flow_evaluator: Some(mock_transform_flow_evaluator),
         mock_dataset_changes: Some(MockDatasetIncrementQueryService::with_increment_between(
-            DatasetIntervalIncrement {
+            MetadataChainIncrementInterval {
                 num_blocks: 1,
                 num_records: 3,
                 updated_watermark: None,
@@ -3533,6 +3536,7 @@ async fn test_derived_dataset_triggered_after_input_change() {
                                 new_head: odf::Multihash::from_digest_sha3_256(b"newest-slice"),
                                 has_more: false,
                             },
+                            data_increment: None,
                         }
                         .into_task_result(),
                     ),
@@ -3696,7 +3700,7 @@ async fn test_derived_dataset_trigger_at_startup_with_external_change_detected()
     let harness = FlowHarness::with_overrides(FlowHarnessOverrides {
         mock_transform_flow_evaluator: Some(mock_transform_flow_evaluator),
         mock_dataset_changes: Some(MockDatasetIncrementQueryService::with_increment_between(
-            DatasetIntervalIncrement {
+            MetadataChainIncrementInterval {
                 num_blocks: 1,
                 num_records: 3,
                 updated_watermark: None,
@@ -3805,6 +3809,7 @@ async fn test_derived_dataset_trigger_at_startup_with_external_change_detected()
                                 new_head: odf::Multihash::from_digest_sha3_256(b"newest-slice"),
                                 has_more: false,
                             },
+                            data_increment: None,
                         }
                         .into_task_result(),
                     ),
@@ -4134,7 +4139,7 @@ async fn test_throttling_derived_dataset_with_2_parents() {
         awaiting_step: Some(Duration::milliseconds(SCHEDULING_ALIGNMENT_MS)), // 10ms,
         mandatory_throttling_period: Some(Duration::milliseconds(SCHEDULING_ALIGNMENT_MS * 10)), /* 100ms */
         mock_dataset_changes: Some(MockDatasetIncrementQueryService::with_increment_between(
-            DatasetIntervalIncrement {
+            MetadataChainIncrementInterval {
                 num_blocks: 2,
                 num_records: 7,
                 updated_watermark: None,
@@ -4225,6 +4230,7 @@ async fn test_throttling_derived_dataset_with_2_parents() {
                                 new_head: odf::Multihash::from_digest_sha3_256(b"foo-new-slice"),
                                 has_more: false,
                             },
+                            data_increment: None,
                         }
                         .into_task_result(),
                     ),
@@ -4254,6 +4260,7 @@ async fn test_throttling_derived_dataset_with_2_parents() {
                                 new_head: odf::Multihash::from_digest_sha3_256(b"fbar-new-slice"),
                                 has_more: false,
                             },
+                            data_increment: None,
                         }
                         .into_task_result(),
                     ),
@@ -4301,6 +4308,7 @@ async fn test_throttling_derived_dataset_with_2_parents() {
                                 new_head: odf::Multihash::from_digest_sha3_256(b"foo-newest-slice"),
                                 has_more: false,
                             },
+                            data_increment: None,
                         }
                         .into_task_result(),
                     ),
@@ -4348,6 +4356,7 @@ async fn test_throttling_derived_dataset_with_2_parents() {
                                 new_head: odf::Multihash::from_digest_sha3_256(b"bar-newest-slice"),
                                 has_more: false,
                             },
+                            data_increment: None,
                         }
                         .into_task_result(),
                     ),
@@ -4720,7 +4729,7 @@ async fn test_batching_condition_records_reached() {
         .times(1)
         .in_sequence(&mut seq_dataset_changes)
         .returning(|_, _, _| {
-            Ok(DatasetIntervalIncrement {
+            Ok(MetadataChainIncrementInterval {
                 num_blocks: 1,
                 num_records: 5,
                 updated_watermark: None,
@@ -4732,7 +4741,7 @@ async fn test_batching_condition_records_reached() {
         .times(1)
         .in_sequence(&mut seq_dataset_changes)
         .returning(|_, _, _| {
-            Ok(DatasetIntervalIncrement {
+            Ok(MetadataChainIncrementInterval {
                 num_blocks: 1,
                 num_records: 7,
                 updated_watermark: None,
@@ -4744,7 +4753,7 @@ async fn test_batching_condition_records_reached() {
         .times(1)
         .in_sequence(&mut seq_dataset_changes)
         .returning(|_, _, _| {
-            Ok(DatasetIntervalIncrement {
+            Ok(MetadataChainIncrementInterval {
                 num_blocks: 2,
                 num_records: 12,
                 updated_watermark: None,
@@ -4757,7 +4766,7 @@ async fn test_batching_condition_records_reached() {
         .times(1)
         .in_sequence(&mut seq_dataset_changes)
         .returning(|_, _, _| {
-            Ok(DatasetIntervalIncrement {
+            Ok(MetadataChainIncrementInterval {
                 num_blocks: 1,
                 num_records: 5,
                 updated_watermark: None,
@@ -4769,7 +4778,7 @@ async fn test_batching_condition_records_reached() {
         .times(1)
         .in_sequence(&mut seq_dataset_changes)
         .returning(|_, _, _| {
-            Ok(DatasetIntervalIncrement {
+            Ok(MetadataChainIncrementInterval {
                 num_blocks: 1,
                 num_records: 5,
                 updated_watermark: None,
@@ -4781,7 +4790,7 @@ async fn test_batching_condition_records_reached() {
         .times(1)
         .in_sequence(&mut seq_dataset_changes)
         .returning(|_, _, _| {
-            Ok(DatasetIntervalIncrement {
+            Ok(MetadataChainIncrementInterval {
                 num_blocks: 2,
                 num_records: 10,
                 updated_watermark: None,
@@ -4867,6 +4876,7 @@ async fn test_batching_condition_records_reached() {
                                 new_head: odf::Multihash::from_digest_sha3_256(b"foo-new-slice"),
                                 has_more: false,
                             },
+                            data_increment: None,
                         }
                         .into_task_result(),
                     ),
@@ -4896,6 +4906,7 @@ async fn test_batching_condition_records_reached() {
                                 new_head: odf::Multihash::from_digest_sha3_256(b"foo-new-slice-2"),
                                 has_more: false,
                             },
+                            data_increment: None,
                         }
                         .into_task_result(),
                     ),
@@ -4925,6 +4936,7 @@ async fn test_batching_condition_records_reached() {
                                 new_head: odf::Multihash::from_digest_sha3_256(b"bar-new-slice"),
                                 has_more: false,
                             },
+                            data_increment: None,
                         }
                         .into_task_result(),
                     ),
@@ -4954,6 +4966,7 @@ async fn test_batching_condition_records_reached() {
                                 new_head: odf::Multihash::from_digest_sha3_256(b"foo-new-slice-3"),
                                 has_more: false,
                             },
+                            data_increment: None,
                         }
                         .into_task_result(),
                     ),
@@ -4983,6 +4996,7 @@ async fn test_batching_condition_records_reached() {
                                 new_head: odf::Multihash::from_digest_sha3_256(b"foo-new-slice-3"),
                                 has_more: false,
                             },
+                            data_increment: None,
                         }
                         .into_task_result(),
                     ),
@@ -5012,6 +5026,7 @@ async fn test_batching_condition_records_reached() {
                                 new_head: odf::Multihash::from_digest_sha3_256(b"bar-new-slice-2"),
                                 has_more: false,
                             },
+                            data_increment: None,
                         }
                         .into_task_result(),
                     ),
@@ -5312,7 +5327,7 @@ async fn test_batching_condition_timeout() {
         .times(1)
         .in_sequence(&mut seq_dataset_changes)
         .returning(|_, _, _| {
-            Ok(DatasetIntervalIncrement {
+            Ok(MetadataChainIncrementInterval {
                 num_blocks: 1,
                 num_records: 5,
                 updated_watermark: None,
@@ -5325,7 +5340,7 @@ async fn test_batching_condition_timeout() {
         .times(1)
         .in_sequence(&mut seq_dataset_changes)
         .returning(|_, _, _| {
-            Ok(DatasetIntervalIncrement {
+            Ok(MetadataChainIncrementInterval {
                 num_blocks: 1,
                 num_records: 3,
                 updated_watermark: None,
@@ -5337,7 +5352,7 @@ async fn test_batching_condition_timeout() {
         .times(1)
         .in_sequence(&mut seq_dataset_changes)
         .returning(|_, _, _| {
-            Ok(DatasetIntervalIncrement {
+            Ok(MetadataChainIncrementInterval {
                 num_blocks: 2,
                 num_records: 8,
                 updated_watermark: None,
@@ -5423,6 +5438,7 @@ async fn test_batching_condition_timeout() {
                                 new_head: odf::Multihash::from_digest_sha3_256(b"foo-new-slice"),
                                 has_more: false,
                             },
+                            data_increment: None,
                         }
                         .into_task_result(),
                     ),
@@ -5452,6 +5468,7 @@ async fn test_batching_condition_timeout() {
                                 new_head: odf::Multihash::from_digest_sha3_256(b"foo-new-slice-2"),
                                 has_more: false,
                             },
+                            data_increment: None,
                         }
                         .into_task_result(),
                     ),
@@ -5483,6 +5500,7 @@ async fn test_batching_condition_timeout() {
                                 new_head: odf::Multihash::from_digest_sha3_256(b"bar-new-slice-2"),
                                 has_more: false,
                             },
+                            data_increment: None,
                         }
                         .into_task_result(),
                     ),
@@ -5630,7 +5648,7 @@ async fn test_batching_condition_watermark() {
         .times(1)
         .in_sequence(&mut seq_dataset_changes)
         .returning(|_, _, _| {
-            Ok(DatasetIntervalIncrement {
+            Ok(MetadataChainIncrementInterval {
                 num_blocks: 1,
                 num_records: 0,
                 updated_watermark: None,
@@ -5643,7 +5661,7 @@ async fn test_batching_condition_watermark() {
         .times(1)
         .in_sequence(&mut seq_dataset_changes)
         .returning(|_, _, _| {
-            Ok(DatasetIntervalIncrement {
+            Ok(MetadataChainIncrementInterval {
                 num_blocks: 1,
                 num_records: 0, // no records, just watermark
                 updated_watermark: Some(Utc::now()),
@@ -5655,7 +5673,7 @@ async fn test_batching_condition_watermark() {
         .times(1)
         .in_sequence(&mut seq_dataset_changes)
         .returning(|_, _, _| {
-            Ok(DatasetIntervalIncrement {
+            Ok(MetadataChainIncrementInterval {
                 num_blocks: 1,
                 num_records: 0, // no records, just watermark
                 updated_watermark: Some(Utc::now()),
@@ -5740,6 +5758,7 @@ async fn test_batching_condition_watermark() {
                                 new_head: odf::Multihash::from_digest_sha3_256(b"foo-new-slice"),
                                 has_more: false,
                             },
+                            data_increment: None,
                         }
                         .into_task_result(),
                     ),
@@ -5769,6 +5788,7 @@ async fn test_batching_condition_watermark() {
                                 new_head: odf::Multihash::from_digest_sha3_256(b"foo-new-slice-2"),
                                 has_more: false,
                             },
+                            data_increment: None,
                         }
                         .into_task_result(),
                     ),
@@ -5800,6 +5820,7 @@ async fn test_batching_condition_watermark() {
                                 new_head: odf::Multihash::from_digest_sha3_256(b"bar-new-slice-2"),
                                 has_more: false,
                             },
+                            data_increment: None,
                         }
                         .into_task_result(),
                     ),
@@ -5947,7 +5968,7 @@ async fn test_batching_condition_with_2_inputs() {
         .times(1)
         .in_sequence(&mut seq_dataset_changes)
         .returning(|_, _, _| {
-            Ok(DatasetIntervalIncrement {
+            Ok(MetadataChainIncrementInterval {
                 num_blocks: 1,
                 num_records: 5,
                 updated_watermark: None,
@@ -5959,7 +5980,7 @@ async fn test_batching_condition_with_2_inputs() {
         .times(1)
         .in_sequence(&mut seq_dataset_changes)
         .returning(|_, _, _| {
-            Ok(DatasetIntervalIncrement {
+            Ok(MetadataChainIncrementInterval {
                 num_blocks: 1,
                 num_records: 5,
                 updated_watermark: None,
@@ -5972,7 +5993,7 @@ async fn test_batching_condition_with_2_inputs() {
         .times(1)
         .in_sequence(&mut seq_dataset_changes)
         .returning(|_, _, _| {
-            Ok(DatasetIntervalIncrement {
+            Ok(MetadataChainIncrementInterval {
                 num_blocks: 1,
                 num_records: 7,
                 updated_watermark: None,
@@ -5984,7 +6005,7 @@ async fn test_batching_condition_with_2_inputs() {
         .times(1)
         .in_sequence(&mut seq_dataset_changes)
         .returning(|_, _, _| {
-            Ok(DatasetIntervalIncrement {
+            Ok(MetadataChainIncrementInterval {
                 num_blocks: 2,
                 num_records: 8,
                 updated_watermark: None,
@@ -5996,7 +6017,7 @@ async fn test_batching_condition_with_2_inputs() {
         .times(1)
         .in_sequence(&mut seq_dataset_changes)
         .returning(|_, _, _| {
-            Ok(DatasetIntervalIncrement {
+            Ok(MetadataChainIncrementInterval {
                 num_blocks: 1,
                 num_records: 7,
                 updated_watermark: None,
@@ -6008,7 +6029,7 @@ async fn test_batching_condition_with_2_inputs() {
         .times(1)
         .in_sequence(&mut seq_dataset_changes)
         .returning(|_, _, _| {
-            Ok(DatasetIntervalIncrement {
+            Ok(MetadataChainIncrementInterval {
                 num_blocks: 1,
                 num_records: 32,
                 updated_watermark: None,
@@ -6110,6 +6131,7 @@ async fn test_batching_condition_with_2_inputs() {
                                 new_head: odf::Multihash::from_digest_sha3_256(b"foo-new-slice"),
                                 has_more: false,
                             },
+                            data_increment: None,
                         }
                         .into_task_result(),
                     ),
@@ -6139,6 +6161,7 @@ async fn test_batching_condition_with_2_inputs() {
                                 new_head: odf::Multihash::from_digest_sha3_256(b"bar-new-slice"),
                                 has_more: false,
                             },
+                            data_increment: None,
                         }
                         .into_task_result(),
                     ),
@@ -6168,6 +6191,7 @@ async fn test_batching_condition_with_2_inputs() {
                                 new_head: odf::Multihash::from_digest_sha3_256(b"foo-new-slice-2"),
                                 has_more: false,
                             },
+                            data_increment: None,
                         }
                         .into_task_result(),
                     ),
@@ -6197,6 +6221,7 @@ async fn test_batching_condition_with_2_inputs() {
                                 new_head: odf::Multihash::from_digest_sha3_256(b"bar-new-slice-2"),
                                 has_more: false,
                             },
+                            data_increment: None,
                         }
                         .into_task_result(),
                     ),
@@ -6226,6 +6251,7 @@ async fn test_batching_condition_with_2_inputs() {
                                 new_head: odf::Multihash::from_digest_sha3_256(b"foo-new-slice-2"),
                                 has_more: false,
                             },
+                            data_increment: None,
                         }
                         .into_task_result(),
                     ),
@@ -6255,6 +6281,7 @@ async fn test_batching_condition_with_2_inputs() {
                                 new_head: odf::Multihash::from_digest_sha3_256(b"baz-new-slice-2"),
                                 has_more: false,
                             },
+                            data_increment: None,
                         }
                         .into_task_result(),
                     ),
@@ -7255,7 +7282,7 @@ async fn test_abort_flow_after_task_finishes() {
         .expect_get_increment_between()
         .times(2)
         .returning(|_, _, _| {
-            Ok(DatasetIntervalIncrement {
+            Ok(MetadataChainIncrementInterval {
                 num_blocks: 1,
                 num_records: 5,
                 updated_watermark: None,
@@ -7350,6 +7377,7 @@ async fn test_abort_flow_after_task_finishes() {
                                 new_head: odf::Multihash::from_digest_sha3_256(b"foo-new-slice"),
                                 has_more: false,
                             },
+                            data_increment: None,
                         }
                         .into_task_result(),
                     ),
@@ -7422,6 +7450,7 @@ async fn test_abort_flow_after_task_finishes() {
                                 new_head: odf::Multihash::from_digest_sha3_256(b"foo-newer-slice"),
                                 has_more: false,
                             },
+                            data_increment: None,
                         }
                         .into_task_result(),
                     ),
@@ -7728,7 +7757,7 @@ async fn test_respect_last_success_time_for_derived_dataset_when_activate_config
         .expect_get_increment_between()
         .times(5)
         .returning(|_, _, _| {
-            Ok(DatasetIntervalIncrement {
+            Ok(MetadataChainIncrementInterval {
                 num_blocks: 1,
                 num_records: 5,
                 updated_watermark: None,
@@ -7837,6 +7866,7 @@ async fn test_respect_last_success_time_for_derived_dataset_when_activate_config
                                 new_head: odf::Multihash::from_digest_sha3_256(b"foo-new-slice"),
                                 has_more: false,
                             },
+                            data_increment: None,
                         }
                         .into_task_result(),
                     ),
@@ -7884,6 +7914,7 @@ async fn test_respect_last_success_time_for_derived_dataset_when_activate_config
                                 new_head: odf::Multihash::from_digest_sha3_256(b"foo-newest-slice"),
                                 has_more: false,
                             },
+                            data_increment: None,
                         }
                         .into_task_result(),
                     ),
@@ -8111,7 +8142,7 @@ async fn test_restart_batching_condition_deadline_on_each_reactivation() {
         .times(3)
         .in_sequence(&mut sequence_dataset_changes)
         .returning(|_, _, _| {
-            Ok(DatasetIntervalIncrement {
+            Ok(MetadataChainIncrementInterval {
                 num_blocks: 1,
                 num_records: 5,
                 updated_watermark: None,
@@ -8123,7 +8154,7 @@ async fn test_restart_batching_condition_deadline_on_each_reactivation() {
         .times(1)
         .in_sequence(&mut sequence_dataset_changes)
         .returning(|_, _, _| {
-            Ok(DatasetIntervalIncrement {
+            Ok(MetadataChainIncrementInterval {
                 num_blocks: 1,
                 num_records: 5,
                 updated_watermark: None,
@@ -8231,6 +8262,7 @@ async fn test_restart_batching_condition_deadline_on_each_reactivation() {
                                 new_head: odf::Multihash::from_digest_sha3_256(b"foo-new-slice"),
                                 has_more: false,
                             },
+                            data_increment: None,
                         }
                         .into_task_result(),
                     ),
@@ -8260,6 +8292,7 @@ async fn test_restart_batching_condition_deadline_on_each_reactivation() {
                                 new_head: odf::Multihash::from_digest_sha3_256(b"bar-new-slice"),
                                 has_more: false,
                             },
+                            data_increment: None,
                         }
                         .into_task_result(),
                     ),
@@ -9734,7 +9767,7 @@ async fn test_dependencies_flow_trigger_instantly_with_zero_batching_rule() {
 
     let harness = FlowHarness::with_overrides(FlowHarnessOverrides {
         mock_dataset_changes: Some(MockDatasetIncrementQueryService::with_increment_between(
-            DatasetIntervalIncrement {
+            MetadataChainIncrementInterval {
                 num_blocks: 1,
                 num_records: 0,
                 updated_watermark: None,
@@ -9823,6 +9856,7 @@ async fn test_dependencies_flow_trigger_instantly_with_zero_batching_rule() {
                                 new_head: odf::Multihash::from_digest_sha3_256(b"newest-slice"),
                                 has_more: false,
                             },
+                            data_increment: None,
                         }
                         .into_task_result(),
                     ),